themes/requirements.txt
community==1.0.0b1
dhlab==2.9.4
ipython==8.5.0
matplotlib==3.6.1
networkx==2.6.3
numpy==1.22.3
pandas==1.4.3
Pillow==9.2.0
python_louvain==0.13
requests==2.27.1
seaborn==0.10.0
streamlit==1.12.2
streamlit_agraph==0.0.35
openpyxl==3.0.10
